chore: move WalletConnectors.ts to bzx-common but haven't Web3ConnectionFactory.ts there. So the one from fulcrum package is used that follows to non-compilable state for explorer, torque and staking
