+++ packages/fulcrum/src/components/AppRouter.tsx
@@ -10,7 +10,7 @@ import { HeaderOps } from '../layout/HeaderOps'
 import { FulcrumProviderEvents } from '../services/events/FulcrumProviderEvents'
 import { ProviderChangedEvent } from '../services/events/ProviderChangedEvent'
 import { FulcrumProvider } from '../services/FulcrumProvider'
-import configProviders from './../config/providers.json'
+import configProviders from 'bzx-common/src/config/providers.ts'
 import { ProviderMenu } from './ProviderMenu'
 import { RiskDisclosure } from './RiskDisclosure'
 import { errors } from 'ethers'
