+++ packages/fulcrum/src/components/ProviderMenu.tsx
@@ -4,7 +4,7 @@ import { ProviderMenuListItem } from './ProviderMenuListItem'
 import { useWeb3React } from '@web3-react/core'
 import { ProviderTypeDictionary } from '../domain/ProviderTypeDictionary'
 import { FulcrumProvider } from '../services/FulcrumProvider'
-import { injected } from '../domain/WalletConnectors'
+import { injected } from 'bzx-common/src/lib/web3ReactUtils'
 import { AbstractConnector } from '@web3-react/abstract-connector'
 import { SwitchButtonInput } from './SwitchButtonInput'
 import Asset from 'bzx-common/src/assets/Asset'
