--- packages/fulcrum/src/config/providers.json
@@ -1,11 +1,0 @@
-{
-  "Alchemy_ApiKey": "DXcS3vM0PJ320XCinLat00Y_UXJNwL_n",
-  "Alchemy_ApiKey_kovan": "lXKYMXYN9O57V8d2jlPH6lqNdS-TAQMp",
-  "Bitski_ClientId": "34cd62b9-f401-4751-adf8-6b3bd4ad7ecc",
-  "Fortmatic_ApiKey": "pk_live_01E9A13083FD8D5F",
-  "Portis_DAppId": "32537ad9-601e-4854-80a2-216dcea0fed6",
-  "Squarelink_ClientId": "1bd995a2f7730d49f889",
-  "Covalent_ApiKey": "ckey_c3bf7f3f22fd465b9afa7650d08",
-  "Google_TrackingID": "GTM-N8H9C42",
-  "Etherscan_Api": "X85FQ7Y6FEZGSUTMKVGATUK5SKFCE76EYA"
-}
