+++ packages/fulcrum/src/domain/WalletHooks.ts
@@ -1,7 +1,7 @@
 import { useState, useEffect } from 'react'
 import { useWeb3React } from '@web3-react/core'
 
-import { injected } from './WalletConnectors'
+import { injected } from 'bzx-common/src/lib/web3ReactUtils'
 
 export function useEagerConnect() {
   const { activate, active } = useWeb3React()
