+++ packages/protocol-explorer/src/components/ProviderMenu.tsx
@@ -7,7 +7,7 @@ import Asset from 'bzx-common/src/assets/Asset'
 import AssetsDictionary from 'bzx-common/src/assets/AssetsDictionary'
 import { ProviderType } from '../domain/ProviderType'
 import ProviderTypeDictionary from '../domain/ProviderTypeDictionary'
-import { injected } from '../domain/WalletConnectors'
+import { injected } from 'bzx-common/src/lib/web3ReactUtils'
 import { ExplorerProvider } from '../services/ExplorerProvider'
 import { ProviderMenuListItem } from './ProviderMenuListItem'
 import { SwitchButtonInput } from './SwitchButtonInput'
