+++ packages/staking-dashboard/src/domain/ProviderTypeDictionary.ts
@@ -26,7 +26,7 @@ import {
   trezor,
   walletconnect,
   walletlink,
-} from './WalletConnectors'
+} from 'bzx-common/src/lib/web3ReactUtils'
 
 const connectorsByName: { [name: string]: AbstractConnector | null } = {
   [ProviderType.MetaMask]: injected,
