+++ packages/staking-dashboard/src/index.tsx
@@ -4,7 +4,7 @@ import ReactDOM from 'react-dom'
 import TagManager from 'react-gtm-module'
 import App from './App'
 import appConfig from 'bzx-common/src/config/appConfig'
-import configProviders from './config/providers.json'
+import configProviders from 'bzx-common/src/config/providers.ts'
 import stakingProvider from './services/StakingProvider'
 import RootStore from './stores/RootStore'
 import './styles/index.scss'
