+++ packages/torque/src/components/ProviderMenu.tsx
@@ -4,7 +4,7 @@ import { ProviderMenuListItem } from './ProviderMenuListItem'
 import { useWeb3React } from '@web3-react/core'
 import { ProviderTypeDictionary } from '../domain/ProviderTypeDictionary'
 import { TorqueProvider } from '../services/TorqueProvider'
-import { injected } from '../domain/WalletConnectors'
+import { injected } from 'bzx-common/src/lib/web3ReactUtils'
 import { AbstractConnector } from '@web3-react/abstract-connector'
 import { SwitchButtonInput } from './SwitchButtonInput'
 import { ReactComponent as CloseIcon } from '../assets/images/ic__close.svg'
