--- packages/torque/src/config/providers.json
@@ -1,12 +1,0 @@
-{
-  "Alchemy_ApiKey": "DXcS3vM0PJ320XCinLat00Y_UXJNwL_n",
-  "Alchemy_ApiKey_kovan": "lXKYMXYN9O57V8d2jlPH6lqNdS-TAQMp",
-  "Bitski_ClientId": "34cd62b9-f401-4751-adf8-6b3bd4ad7ecc",
-  "Fortmatic_ApiKey": "pk_live_01E9A13083FD8D5F",
-  "Portis_DAppId": "07378a1d-4793-45d7-a25e-23c1681d696b",
-  "Squarelink_ClientId": "1bd995a2f7730d49f889",
-  "Covalent_ApiKey": "ckey_c3bf7f3f22fd465b9afa7650d08",
-  "Google_TrackingID": "GTM-NQCQR39",
-  "Infura_ApiKey": "7989ee6b11324cc49f18b8ab7be5a7c4",
-  "Etherscan_Api": "X85FQ7Y6FEZGSUTMKVGATUK5SKFCE76EYA"
-}
