+++ packages/torque/src/services/TorqueProvider.ts
@@ -57,7 +57,7 @@ import { RequestStatus } from '../domain/RequestStatus'
 import { RequestTask } from '../domain/RequestTask'
 import { TasksQueueEvents } from './events/TasksQueueEvents'
 import { TasksQueue } from './TasksQueue'
-import configProviders from '../config/providers.json'
+import configProviders from 'bzx-common/src/config/providers.ts'
 import { LiquidationEvent } from '../domain/events/LiquidationEvent'
 
 const isMainnetProd =
